Propose OpenEBS into CNCF Sandbox
This PR is to propose OpenEBS into CNCF Sandbox.

Here is a link to the presentation we gave about OpenEBS
to the CNCF Storage WG.
https://docs.google.com/presentation/d/1MEx9EyvFW1cLZhFxUvWVd4bzoRS7PNnMqgmbIZaoulw/edit

OpenEBS Community is pretty excited at the prospect of being in
 CNCF Sandbox and so are some of the users we reached out to with
 this idea in the past 2 weeks. All of them saw this is a logical
 next step for an always open source  project like OpenEBS.

Also we have some traditional storage vendors and cloud service providers
as well that will be more able to contribute once OpenEBS is in the CNCF
as opposed to being seen as a one vendor project.

Signed-off-by: kmova <kiran.mova@openebs.io>