Propose spiffe and spire for graduation (#778)
* Propose spiffe and spire for graduation

The SPIFFE and SPIRE projects were started with an ambitious vision: to automatically provide every workloads in a distributed system with an identity that is cryptographically verifiable, automatically rotated, interoperable and portable. The goal is an interoperable workload identity format (SPIFFE), and supporting tooling to issue those identities (SPIRE), that allows components in a distributed system to ensure the authenticity, integrity and confidentiality of messages passed between them. 

The growth of the SPIFFE and SPIRE projects, with support from the CNCF as it moved through the Sandbox and Incubation phases, is evidenced in many dimensions ranging from maturity and robustness to improved user experience and broader reach. 

We believe the projects meet the criteria for graduation and are currently looking for a TOC sponsor.

Signed-off-by: Andres Vega <andresv@vmware.com>

* fixes capitalization

Signed-off-by: Andres Vega <andresv@vmware.com>

* adds links and slight tweak to stats

adds links to book and katacoda

the stats are fairly fluid and the numbers captured are a close approximation at the time of this pr

* expand on adopters

* Update spiffe_spire.md

Signed-off-by: Andres Vega <andresv@vmware.com>