Update longhorn.adoc
 Hi ,
I have made a small correction to address a typo in the code. This pull request aims to fix the error and improve the overall code quality.
Please review the commit and consider merging it into the main branch. Feel free to provide any feedback or suggestions for improvement.

Thank you for maintaining this open source project, and I appreciate the opportunity to contribute! I'm learning kubernetes so that I can contribute into larger problems. It would be helpful if you can share some resources with me for learning.

Signed-off-by: Samadrita-Bhattacharya <108999524+Samadrita-Bhattacharya@users.noreply.github.com>