working groups: add process for creating new WGs in a TAG
The whole process needs to be revamped as part of https://github.com/cncf/toc/issues/1158.
There was some work that was started as part of https://github.com/cncf/toc/pull/868,
but the PR has languished.

This PR keeps the existing process and defines a lightweight process for
creation of new WGs in a TAG. This mirrors the process we follow today.

Signed-off-by: Nikhita Raghunath <nikitaraghunath@gmail.com>