Add new project level templates
This is a templated version of the items discussed from the project-moving-levels task force. It's split into three documents, one for each level. I added a number so they display in order but ideally there's a github issue template that will just link to these directly. 

The idea is a project files a tracking issue, and then PRs in one of these and then that becomes the living document of the lifecycle of the project for that level. There are placeholders for the TOC to be able to fill in the evaluation for each criteria bullet. Happy to make adjustments and fixes depending on feedback!






Signed-off-by: Jorge O. Castro <jorge.castro@gmail.com>