Add Projects Team sync with a project moving a level
This seems like the right place for this but I can move it wherever. Basically when a project graduates we'd like to meet with them, get the congratulations out of the way, and then go over all of the project benefits at that level, answer questions about their benefits, that sort of thing. 

This will include a rebrief from the events team or a representative so projects understand how KubeCon works from a logistical perspective. 

Signed-off-by: Jorge O. Castro <jorge.castro@gmail.com>